+++ apps/vscode/src/providers/assist/render-assist.ts
@@ -31,7 +31,6 @@ import {
   SignatureInformation,
 } from "vscode";
 import { escapeRegExpCharacters } from "core";
-import { isQuartoDoc, isQuartoYaml } from "../../core/doc";
 
 const kAssistHelp = "Quarto: Help";
 const kAssistEquation = "Quarto: Equation";
@@ -81,11 +80,6 @@ export async function renderActiveAssist(
   // must have an active text editor
   const editor = window.activeTextEditor;
   if (!editor) {
-    return undefined;
-  }
-
-  // must be a quarto document
-  if (!isQuartoDoc(editor.document) && !isQuartoYaml(editor.document)) {
     return undefined;
   }