+++ apps/vscode/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Toggle comment (%%) for Mermaid diagram editing
 - Support math preview in list items with > 2 blocks
 - Don't show GitLens hovers in Quarto Assist panel
+- Restore Quarto Assist for non-qmd file types
 
 ## 1.83.0 (Release on 28 April 2023)